refactor: 두 alert modifier를 별도 파일로 분리
